Reduce contention on level data map so that map loads/unloads are more independent
Store level data by shared (threadsafe) pointer instead of by value so we can only lock the owning map more briefly. This means that we're not locking the map as a whole while we load / write / unload the content of one level so they can happen in parallel. Updating the map itself is still (briefly) locked as are individual instances of level data, but the scope is much narrower.